refactor: 모듈위치, 미사용 코드 제거
