+++ build.gradle
@@ -81,6 +81,9 @@ project(":layer-api") {
 		// jpa
 		implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 
+		//Swagger
+		implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2'
+
 		runtimeOnly 'com.mysql:mysql-connector-j'
 	}
 
@@ -94,6 +97,7 @@ project(":layer-common") {
 
 	dependencies {
 		implementation 'org.springframework.boot:spring-boot-starter-web'
+		implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	}
 
 }
