--- layer-api/src/main/java/org/layer/api/controller/HelloController.java
@@ -1,14 +1,0 @@
-package org.layer.api.controller;
-
-import lombok.RequiredArgsConstructor;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/test")
-public class HelloController {
-
-}
