--- layer-api/src/main/java/org/layer/auth/dto/controller/SignInRequest.java
@@ -1,7 +1,0 @@
-package org.layer.auth.dto.controller;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import org.layer.domain.member.entity.SocialType;
-
-public record SignInRequest(@JsonProperty("social_type") SocialType socialType) {
-}
