--- layer-api/src/main/java/org/layer/auth/dto/controller/SignUpRequest.java
@@ -1,9 +1,0 @@
-package org.layer.auth.dto.controller;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import org.layer.domain.member.entity.SocialType;
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record SignUpRequest(SocialType socialType, String name) {
-}
