--- layer-api/src/main/java/org/layer/auth/dto/service/ReissueTokenServiceResponse.java
@@ -1,10 +1,0 @@
-package org.layer.auth.dto.service;
-
-import org.layer.auth.jwt.JwtToken;
-import org.layer.domain.member.entity.Member;
-
-public record ReissueTokenServiceResponse(Long memberId, JwtToken jwtToken) {
-    public static ReissueTokenServiceResponse of(Member member, JwtToken jwtToken) {
-        return new ReissueTokenServiceResponse(member.getId(), jwtToken);
-    }
-}
