--- layer-api/src/main/java/org/layer/auth/exception/TokenException.java
@@ -1,5 +1,0 @@
-package org.layer.auth.exception;
-
-// 임시로 만들어놓은 Exception입니다
-public class TokenException extends RuntimeException {
-}
