--- layer-api/src/main/java/org/layer/auth/jwt/JwtToken.java
@@ -1,16 +1,0 @@
-package org.layer.auth.jwt;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class JwtToken {
-    private final String accessToken;
-    private final String refreshToken;
-
-    @Builder
-    public JwtToken(String accessToken, String refreshToken) {
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-    }
-}
