--- layer-api/src/main/java/org/layer/auth/jwt/JwtValidationType.java
@@ -1,7 +1,0 @@
-package org.layer.auth.jwt;
-
-public enum JwtValidationType {
-    VALID_JWT,
-    INVALID_JWT;
-}
-
