+++ layer-api/src/main/java/org/layer/common/exception/GlobalExceptionHandler.java
@@ -1,8 +1,6 @@
 package org.layer.common.exception;
 
 import lombok.extern.slf4j.Slf4j;
-import org.layer.common.exception.BaseCustomException;
-import org.layer.common.exception.ExceptionType;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
