+++ layer-api/src/main/java/org/layer/config/SecurityConfig.java
@@ -2,7 +2,7 @@ package org.layer.config;
 
 
 import lombok.RequiredArgsConstructor;
-import org.layer.auth.jwt.JwtAuthenticationFilter;
+import org.layer.domain.jwt.JwtAuthenticationFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -40,6 +40,7 @@ public class SecurityConfig {
                                 .requestMatchers(new AntPathRequestMatcher("/api/auth/sign-up")).permitAll()
                                 .requestMatchers(new AntPathRequestMatcher("/api/auth/oauth2/google/code")).permitAll()
                                 .requestMatchers(new AntPathRequestMatcher("/api/auth/oauth2/google")).permitAll()
+                                .requestMatchers(new AntPathRequestMatcher("/api/test")).permitAll()
                                 .anyRequest().authenticated()
                 );
     }
@@ -51,7 +52,4 @@ public class SecurityConfig {
                 .requestMatchers(new AntPathRequestMatcher("/swagger-ui/**")).permitAll()
                 .requestMatchers(new AntPathRequestMatcher("/docs/**")).permitAll());
     }
-
-
-
 }
