+++ layer-api/src/main/java/org/layer/domain/auth/controller/dto/SignInRequest.java
@@ -1,0 +1,7 @@
+package org.layer.domain.auth.controller.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import org.layer.domain.member.entity.SocialType;
+
+public record SignInRequest(@JsonProperty("social_type") SocialType socialType) {
+}
