+++ layer-api/src/main/java/org/layer/domain/auth/controller/dto/SignUpRequest.java
@@ -1,0 +1,9 @@
+package org.layer.domain.auth.controller.dto;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import org.layer.domain.member.entity.SocialType;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public record SignUpRequest(SocialType socialType, String name) {
+}
