+++ layer-api/src/main/java/org/layer/domain/auth/controller/dto/SignUpResponse.java
@@ -1,0 +1,21 @@
+package org.layer.domain.auth.controller.dto;
+
+import org.layer.domain.auth.service.dto.SignUpServiceResponse;
+import org.layer.domain.member.entity.MemberRole;
+import org.layer.domain.member.entity.SocialType;
+
+public record SignUpResponse(Long memberId,
+                                    String name,
+                                    String email,
+                                    MemberRole memberRole,
+                                    String SocialId,
+                                    SocialType socialType) {
+    public static SignUpResponse of(SignUpServiceResponse signUpServiceResponse) {
+        return new SignUpResponse(signUpServiceResponse.memberId(),
+                signUpServiceResponse.name(),
+                signUpServiceResponse.email(),
+                signUpServiceResponse.memberRole(),
+                signUpServiceResponse.socialId(),
+                signUpServiceResponse.socialType());
+    }
+}
