+++ layer-api/src/main/java/org/layer/domain/auth/exception/AuthException.java
@@ -1,0 +1,5 @@
+package org.layer.domain.auth.exception;
+
+// OAuth를 위해 임시로 만들어놓은 Exception입니다.
+public class AuthException extends RuntimeException {
+}
