+++ layer-api/src/main/java/org/layer/domain/jwt/JwtToken.java
@@ -1,0 +1,13 @@
+package org.layer.domain.jwt;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class JwtToken {
+    private final String accessToken;
+    private final String refreshToken;
+}
