+++ layer-api/src/main/java/org/layer/domain/jwt/JwtValidationType.java
@@ -1,0 +1,7 @@
+package org.layer.domain.jwt;
+
+public enum JwtValidationType {
+    VALID_JWT,
+    INVALID_JWT
+}
+
