+++ layer-api/src/main/java/org/layer/domain/jwt/SecretKeyFactory.java
@@ -1,0 +1,22 @@
+package org.layer.domain.jwt;
+
+import io.jsonwebtoken.security.Keys;
+import lombok.RequiredArgsConstructor;
+import org.layer.config.AuthValueConfig;
+import org.springframework.stereotype.Component;
+
+import javax.crypto.SecretKey;
+
+import static java.util.Base64.getEncoder;
+
+@RequiredArgsConstructor
+@Component
+public class SecretKeyFactory {
+    private final AuthValueConfig authValueConfig;
+
+    SecretKey createSecretKey() {
+        String encodedKey = getEncoder().encodeToString(authValueConfig.getJWT_SECRET().getBytes());
+        return Keys.hmacShaKeyFor(encodedKey.getBytes());
+    }
+
+}
